requirements.txt
asgiref==3.7.2
Django==4.2.9
sqlparse==0.4.4
typing_extensions==4.9.0
djangorestframework==3.14.0